Updated the format of the location xmls again, changing how call quantity is factored, and also time of day when calls are created more often than others.

+++ AgencyDispatchFramework/Simulation/RegionCrimeInfo.cs
@@ -10,7 +10,7 @@
         /// <summary>
         /// Gets the average crime level index of this Region
         /// </summary>
-        public int AverageCrimeCalls { get; set; }
+        public double AverageCrimeCalls { get; set; }
 
         /// <summary>
         /// Gets the minimum amout of calls to expect from this region
